+++ project/3rdparty/include/live555/liveMedia/liveMedia_version.hh
@@ -19,8 +19,8 @@ along with this library; if not, write to the Free Software Foundation, Inc.,
 #ifndef _LIVEMEDIA_VERSION_HH
 #define _LIVEMEDIA_VERSION_HH
 
-#define LIVEMEDIA_LIBRARY_VERSION_STRING	"2024.04.14"
-#define LIVEMEDIA_LIBRARY_VERSION_INT		1713052800
+#define LIVEMEDIA_LIBRARY_VERSION_STRING	"2024.04.19"
+#define LIVEMEDIA_LIBRARY_VERSION_INT		1713484800
 
 extern char const* const liveMediaLibraryVersionStr;
 extern int const liveMediaLibraryVersionInt;